MINOR: Remove misleading mxEvent.CHANGE because it is undefined
mxUtils.CHANGE is undefined (as opposed to mxEvent.CHANGE), so the
listener that is attached to ui.editor does not do what you think!

Now, mxUtils.CHANGE being undefined, the handler passed to
addListener() gets triggered for all possible events. The patch
does not change anything but clears an ambiguity in the code.