MAJOR: Rework plugin initialization to add support for embed mode
We want to support draw.io in embed mode because we are building an
integration to be able to edit and save diagrams in Kompot. The main
issue is that embed mode does not fire the "fileLoaded" event, which
is the base for initializing the plugin and the ui. The patch
simulates that event by intercepting the "load" event sent by the
embed mode to the parent window.