MEDIUM: Build live panel as later as possible to reduce timing issues
When switching from one diagram to another without relaoding the ui,
I encountered some cases were the live tab click did not do anything,
especially in embed mode. There may be timing issue when the page root
changes. Building the panel depends on identification of selected node
or page root id, it makes sense to do it on the fly when the panel is
going to be displayed.