BUG/MEDIUM: live: Fix lightbox fit and resize on graph load
First live data fetch performed during startScheduleUpdate() loop
is sychronous, ending up with calls to ui.editor.setGraphXml()
from upgradeCellLiveNode(). This probably disturb lightbox fit and
resize on graph load, so let's call it here. It is good to call
it as well on page change to rescale the new graph.